+++ firmware/hdl/packages/gt_mp7_core_pkg.vhd
@@ -1,15 +1,19 @@
+-- Description:
+-- Package for definitions of ugt core.
 
 -- actual versions:
 -- use "FRAME_VERSION" as mp7_ugt release fw version (used for tag name).
--- mp7_ugt (=FRAME_VERSION): v1.22.0
+-- mp7_ugt (=FRAME_VERSION): v1.21.4
 
 -- use "GT_VERSION" as mp7_ugt release fw version (used for tag name).
--- gt: v1.22.0
--- frame: v1.3.0 (see frame.vhd)
+-- gt: v1.21.4
+-- frame: v1.3.1 (see frame.vhd)
 -- gtl: v1.18.0 (see gtl_module_tpl.vhd)
--- fdl: v1.4.0 (see fdl_module.vhd)
+-- fdl: v1.4.1 (see fdl_module.vhd)
 
--- HB 2022-05-31: v1.21.0 - Signal start_sync_bc0_int used for reset of prescale counter (instead of begin_lumi_section). Removed sres signals for counters in FDL, not used anymore.
+-- HB 2022-09-26: v1.21.3 - Added script run_latex_w_versions.py (extract versions from gt_mp7_core_pkg.vhd, creates versions.tex and run latex ['make']).
+-- HB 2022-09-20: v1.21.1 - Updated scripts for MP7 patch in run_synth_ipbb.py.
+-- HB 2022-08-16: v1.21.0 - Signal start_sync_bc0_int used for reset of prescale counter (instead of begin_lumi_section). Removed sres signals for counters in FDL, not used anymore.
 -- HB 2021-05-31: v1.20.2 - Updated script run_simulation_questa.py (removed UGT_GITLAB_PWD)
 -- HB 2021-04-25: v1.20.1 - Updated and added scripts.
 -- HB 2021-02-03: v1.20.0 - "GT_VERSION" is (used in frame_module_info.vhd) in register of "FRAME_VERSION". "FRAME_VERSION" is now in register of "SVN_REVISION_NUMBER".
@@ -25,7 +29,10 @@
 -- v1.15.2: Added tcl script for "manualy" bit file generation (after timing errors)
 --
 -- gtl history:
+
 -- HB 2022-09-12: v1.18.0: Module for "anomaly detection trigger (ADT)" test.
+-- HB 2022-09-23: v1.17.4: Used "delay_pipeline" for condition output in esums_conditions.vhd, min_bias_hf_conditions.vhd and towercount_condition.vhd.
+-- HB 2022-09-02: v1.17.3: Cleaned up.
 -- HB 2022-02-16: v1.17.2: Bug fixed in comb_conditions.vhd.
 -- HB 2021-12-09: v1.17.1: Updated logic for jet DISP cut.
 -- HB 2021-10-23: v1.17.0: Implemented logic for jet DISP (displaced) cut [DISP = bit 27 of jet data] in calo_comparators.vhd. Bug fixed in correlation_cuts_wrapper.vhd.
@@ -76,8 +83,9 @@ package gt_mp7_core_pkg is
 -- ==================================================================================================
 -- GT firmware version
     constant GT_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant GT_MINOR_VERSION      : integer range 0 to 255 := 22;
-    constant GT_REV_VERSION        : integer range 0 to 255 := 0;
+    constant GT_MINOR_VERSION      : integer range 0 to 255 := 21;
+    constant GT_REV_VERSION        : integer range 0 to 255 := 4;
+
 	constant GT_VERSION : std_logic_vector(31 downto 0) := X"00" &
            std_logic_vector(to_unsigned(GT_MAJOR_VERSION, 8)) &
            std_logic_vector(to_unsigned(GT_MINOR_VERSION, 8)) &
@@ -85,7 +93,7 @@ package gt_mp7_core_pkg is
 -- FRAME firmware version
     constant FRAME_MAJOR_VERSION      : integer range 0 to 255 := 1;
     constant FRAME_MINOR_VERSION      : integer range 0 to 255 := 3;
-    constant FRAME_REV_VERSION        : integer range 0 to 255 := 0;
+    constant FRAME_REV_VERSION        : integer range 0 to 255 := 1;
 -- GTL firmware (fix part) version
     constant GTL_FW_MAJOR_VERSION      : integer range 0 to 255 := 1;
     constant GTL_FW_MINOR_VERSION      : integer range 0 to 255 := 18;
@@ -93,7 +101,8 @@ package gt_mp7_core_pkg is
 -- FDL firmware version
     constant FDL_FW_MAJOR_VERSION      : integer range 0 to 255 := 1;
     constant FDL_FW_MINOR_VERSION      : integer range 0 to 255 := 4;
-    constant FDL_FW_REV_VERSION        : integer range 0 to 255 := 0;
+    constant FDL_FW_REV_VERSION        : integer range 0 to 255 := 1;
+
 -- ==================================================================================================
 
     constant BUNCHES_PER_ORBIT : natural range 3564 to 3564 := LHC_BUNCH_COUNT;
@@ -107,7 +116,6 @@ package gt_mp7_core_pkg is
 
 -- HB 2014-07-08: ipbus_rst is high active, RST_ACT changed to '1' (for lhc_rst [in gt_mp7_core_pkg.vhd]) to get proper reset-conditions,
 --                because in delay_line_sl.vhd and delay_line_slv.vhd both resets are used !!!
--- constant RST_ACT : std_logic := '0';
     constant RST_ACT : std_logic := '1';
     constant RST_ACT_ROP : std_logic := '0';