+++ firmware/hdl/payload/gtl/common/calo_comparators.vhd
@@ -3,6 +3,7 @@
 -- Comparators for energy, pseudorapidity, azimuth angle and isolation of calo objects
 
 -- Version history:
+-- HB 2022-09-06: cleaned up.
 -- HB 2021-12-09: updated DISP logic for jets.
 -- HB 2021-10-19: added DISP cut for jets.
 -- HB 2021-10-18: bug fix (in jet_phi_windows_comp_i and tau_phi_windows_comp_i).
@@ -17,8 +18,6 @@
 
 library ieee;
 use ieee.std_logic_1164.all;
-use ieee.std_logic_arith.all;
-use ieee.std_logic_unsigned.all; -- for et and phi comparators
 
 use work.gtl_pkg.all;
 
@@ -119,7 +118,7 @@ begin
                 phi_w2_lower_limit => phi_w2_lower_limit(EG_PHI_HIGH-EG_PHI_LOW downto 0)
             )
             port map(
-                phi => data_i(EG_PHI_HIGH downto EG_PHI_LOW),
+                phi => phi(EG_PHI_HIGH-EG_PHI_LOW downto 0),
                 phi_comp_o => phi_comp
             );
 
@@ -161,7 +160,7 @@ begin
                 phi_w2_lower_limit => phi_w2_lower_limit(JET_PHI_HIGH-JET_PHI_LOW downto 0)
             )
             port map(
-                phi => data_i(JET_PHI_HIGH downto JET_PHI_LOW),
+                phi => phi(JET_PHI_HIGH-JET_PHI_LOW downto 0),
                 phi_comp_o => phi_comp
             );
 
@@ -203,7 +202,7 @@ begin
                 phi_w2_lower_limit => phi_w2_lower_limit(TAU_PHI_HIGH-TAU_PHI_LOW downto 0)
             )
             port map(
-                phi => data_i(TAU_PHI_HIGH downto TAU_PHI_LOW),
+                phi => phi(TAU_PHI_HIGH-TAU_PHI_LOW downto 0),
                 phi_comp_o => phi_comp
             );