+++ firmware/hdl/payload/gtl/common/esums_4_corr_cond.vhd
@@ -1,14 +1,13 @@
 
 -- Description:
--- Correlation cuts module
+-- Esums for correlation cuts.
 
 -- Version history:
+-- HB 2022-09-06: cleaned up.
 -- HB 2021-02-12: first design.
 
 library ieee;
 use ieee.std_logic_1164.all;
-use ieee.std_logic_unsigned.all;
-use ieee.std_logic_arith.all;
 
 use work.gtl_pkg.all;