+++ firmware/hdl/payload/gtl/common/esums_comparators.vhd
@@ -1,8 +1,9 @@
 
--- Desription:
--- Comparators fro esums
+-- Description:
+-- Comparators for esums
 
 -- Version history:
+-- HB 2022-09-06: cleaned up.
 -- HB 2021-05-18: changed slice parameter.
 -- HB 2020-12-14: changed "phi cuts", used "nr_phi_windows" now. New logic.
 -- HB 2020-01-30: removed "no_esums"
@@ -13,8 +14,7 @@
 
 library ieee;
 use ieee.std_logic_1164.all;
-use ieee.std_logic_arith.all;
-use ieee.std_logic_unsigned.all; -- for function "CONV_INTEGER"
+use ieee.std_logic_unsigned.all;
 
 use work.gtl_pkg.all;