+++ firmware/hdl/payload/gtl_module_tpl.vhd
@@ -3,6 +3,8 @@
 
 -- Version-history:
 -- HB 2022-09-12: v1.18.0: Module for "anomaly detection trigger (ADT)" test.
+-- HB 2022-09-23: v1.17.4: Used "delay_pipeline" for condition output in esums_conditions.vhd, min_bias_hf_conditions.vhd and towercount_condition.vhd.
+-- HB 2022-09-05: v1.17.3: Cleaned up.
 -- HB 2022-02-16: v1.17.2: Bug fixed in comb_conditions.vhd.
 -- HB 2021-10-23: v1.17.1: Updated logic for jet DISP cut.
 -- HB 2021-10-23: v1.17.0: Implemented logic for jet DISP (displaced) cut [DISP = bit 27 of jet data] in calo_comparators.vhd. Bug fixed in correlation_cuts_wrapper.vhd.
@@ -47,8 +49,6 @@
 
 library ieee;
 use ieee.std_logic_1164.all;
-use ieee.std_logic_arith.all;
-use ieee.std_logic_unsigned.all;
 
 use work.fdl_pkg.all;
 use work.gtl_pkg.all;