+++ CHANGELOG.md
@@ -4,15 +4,59 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [v1.22.0] - 2022-09-12
+## [v1.21.4] - 2022-010-10
 ### Comment
 
-- mp7_ugt firmware release v1.22.0
+- mp7_ugt firmware release v1.21.4
 - test version for "anomaly detection trigger (ADT)"
 
 ### Changed
 - source files:
   - ../payload/gtl_module_tpl.vhd
+
+## [v1.21.3] - 2022-09-29
+### Comment
+
+- mp7_ugt firmware release v1.21.3
+- updated scripts for MP7 patch in run_synth_ipbb.py and run_simulation_questa.py
+
+### Changed
+- script files:
+  - scripts/mp7patch.py
+  - scripts/run_synth_ipbb.py
+  - scripts/run_simulation_questa.py
+- simulation file:
+  - ../sim/scripts/templates/gtl_fdl_wrapper_tpl_questa.do
+- source files:
+  - ../packages/gt_mp7_core_pkg.vhd
+
+### Added
+- source files:
+  - ../packages/gt_mp7_top_pkg_sim.vhd
+
+## [v1.21.2] - 2022-09-28
+### Comment
+
+- mp7_ugt firmware release v1.21.2
+- cleaned up VHDL code
+
+### Changed
+- multiple VHDL files
+
+## [v1.21.1] - 2022-09-26
+### Comment
+
+- mp7_ugt firmware release v1.21.1
+- added script run_latex_w_versions.py (extract versions from gt_mp7_core_pkg.vhd, creates versions.tex and run latex ['make'])
+
+### Added
+- script files:
+  - scripts/run_latex_w_versions.py
+
+### Changed
+- source files:
+  - ../packages/gt_mp7_core_pkg.vhd
+>>>>>>> master
 
 ## [v1.21.0] - 2022-09-14
 ### Comment